categorization-connector/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
httpx==0.25.1
dnspython==2.4.2
redis==5.0.1
prometheus-client==0.19.0
python-dotenv==1.0.0
pyyaml==6.0.1
hvac==1.2.1
pydantic==2.5.2
pydantic-settings==2.1.0
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
respx==0.20.2
